move dash virtual objects in the correct folder
